chore: revert changes to handwrittenvalidation
